Fix tests with clang by ensuring that all 64 bits of unused rrcall syscall arguments are zero

+++ src/test/util_internal.h
@@ -7,16 +7,19 @@
 #include "rrcalls.h"
 
 int running_under_rr(void) {
-  return 0 == syscall(SYS_rrcall_check_presence, 0, 0, 0, 0, 0, 0);
+  return 0 == syscall(SYS_rrcall_check_presence, (long)0, (long)0,
+      (long)0, (long)0, (long)0, (long)0);
 }
 
 void rr_detach_teleport(void) {
-  int err = syscall(SYS_rrcall_detach_teleport, 0, 0, 0, 0, 0, 0);
+  int err = syscall(SYS_rrcall_detach_teleport, (long)0, (long)0,
+      (long)0, (long)0, (long)0, (long)0);
   test_assert(err == 0);
 }
 
 void rr_freeze_tid(pid_t tid, int freeze) {
-  int err = syscall(SYS_rrcall_freeze_tid, tid, freeze, 0, 0, 0, 0);
+  int err = syscall(SYS_rrcall_freeze_tid, tid, freeze,
+      (long)0, (long)0, (long)0, (long)0);
   test_assert(err == 0 && "Failed to freeze tid");
 }
